LayoutTests/transforms/2d/transform-value-types-expected.txt

PASS jsWrapperClass(transformRule) is 'CSSValueList'
PASS jsWrapperClass(transformRule.__proto__) is 'CSSValueListPrototype'
PASS jsWrapperClass(transformRule.constructor) is 'Function'
PASS jsWrapperClass(transformRule[0]) is 'WebKitCSSTransformValue'
PASS jsWrapperClass(transformRule[0].__proto__) is 'WebKitCSSTransformValuePrototype'
PASS jsWrapperClass(transformRule[0].constructor) is 'Function'
PASS transformRule[0].operationType is WebKitCSSTransformValue.CSS_TRANSLATE
PASS transformRule[0].cssText is 'translate(10px)'
PASS transformRule[1].operationType is WebKitCSSTransformValue.CSS_TRANSLATEX
PASS transformRule[1].cssText is 'translateX(10px)'
PASS transformRule[2].operationType is WebKitCSSTransformValue.CSS_TRANSLATEY
PASS transformRule[2].cssText is 'translateY(10px)'
PASS transformRule[3].operationType is WebKitCSSTransformValue.CSS_ROTATE
PASS transformRule[3].cssText is 'rotate(10deg)'
PASS transformRule[4].operationType is WebKitCSSTransformValue.CSS_ROTATE
PASS transformRule[4].cssText is 'rotate(11rad)'
PASS transformRule[5].operationType is WebKitCSSTransformValue.CSS_ROTATE
PASS transformRule[5].cssText is 'rotate(12grad)'
PASS transformRule[6].operationType is WebKitCSSTransformValue.CSS_ROTATE
PASS transformRule[6].cssText is 'rotate(13turn)'
PASS transformRule[7].operationType is WebKitCSSTransformValue.CSS_SCALE
PASS transformRule[7].cssText is 'scale(2)'
PASS transformRule[8].operationType is WebKitCSSTransformValue.CSS_SCALEX
PASS transformRule[8].cssText is 'scaleX(1.2)'
PASS transformRule[9].operationType is WebKitCSSTransformValue.CSS_SCALEY
PASS transformRule[9].cssText is 'scaleY(1.2)'
PASS transformRule[10].operationType is WebKitCSSTransformValue.CSS_SKEW
PASS transformRule[10].cssText is 'skew(10deg, 10rad)'
PASS transformRule[11].operationType is WebKitCSSTransformValue.CSS_SKEWX
PASS transformRule[11].cssText is 'skewX(10deg)'
PASS transformRule[12].operationType is WebKitCSSTransformValue.CSS_SKEWY
PASS transformRule[12].cssText is 'skewY(11grad)'
PASS transformRule[13].operationType is WebKitCSSTransformValue.CSS_MATRIX
PASS transformRule[13].cssText is 'matrix(1, 0, 0, 1, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
